Better currency handling if invalid currency is selected
